fix: functional audit remediation (F-001 through F-008)
Resolves 7 findings from the 2026-03-10 functional audit:

- F-001: Budget version dataSource uses per-definition value (MANUAL for
  budget, IMPORTED for actuals) instead of hardcoded IMPORTED, unblocking
  revenue calculation
- F-002: Fee grid fixtures include term1/term2/term3 splits; normalize
  nationality codes (FR→Francais, NAT→Nationaux, AUT→Autres) and tariff
  codes (PLEIN→Plein, R3P→R3+) across all fixture files
- F-003: Align GRADE_CODES and GRADE_PROGRESSION with DB seed data
  (6EME/5EME/4EME/3EME/2NDE/1ERE/TERM), fixing AY2 cohort lookups
- F-004: Seed nationality_breakdown table from enrollment detail during
  migration with computed weights
- F-005: Rewrite URL-to-store sync to re-read on every location.search
  change instead of once on mount
- F-006: Discount importer nationality handling consistent with Prisma
  compound unique constraint
- F-008: Revenue parser preserves all 12 months (no longer drops month 7)

All 818 tests pass. Zero type errors. Zero lint violations.

+++ apps/api/src/migration/importers/employees.ts
@@ -59,7 +59,7 @@ export async function importEmployees(
 				) VALUES (
 					$1, $2, $3, $4, $5,
 					$6, $7, $8,
-					$9, $10, $11, $12,
+					$9, $10, $11, $12::numeric,
 					pgp_sym_encrypt($13, $14),
 					pgp_sym_encrypt($15, $14),
 					pgp_sym_encrypt($16, $14),
@@ -67,7 +67,7 @@ export async function importEmployees(
 					pgp_sym_encrypt($18, $14),
 					pgp_sym_encrypt($19, $14),
 					$20,
-					$21, $22,
+					$21::numeric, $22::numeric,
 					$23, NOW(), NOW()
 				)
 				ON CONFLICT (version_id, employee_code)
